+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+- **PETRUSH_PS1 escapes** `\w` `\u` `\h` `\n` `\$` `\\` (UX-15). Default continua `petrush> `.
+
 ## [0.3.1] - 2026-08-03 - pushd/popd + history bangs
 
 ### Added